pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "certmesh"
version = "3.0.0"
description = "Automated TLS certificate lifecycle client for DigiCert, Venafi TPP, HashiCorp Vault PKI, and AWS ACM"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = [
    "tls",
    "certificate",
    "digicert",
    "venafi",
    "vault",
    "pki",
    "acm",
    "aws",
    "automation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]

dependencies = [
    "requests>=2.31.0",
    "tenacity>=8.2.0",
    "hvac>=2.1.0",
    "cryptography>=41.0.0",
    "PyYAML>=6.0.1",
    "boto3>=1.28.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=22.0.0",
    "python-jose[cryptography]>=3.3.0",
    "prometheus-client>=0.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "slowapi>=0.1.9",
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
    "freezegun>=1.3.0",
    "moto[acm,acm-pca,iam,sts,secretsmanager]>=4.2.0",
    "ruff>=0.4.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
certmesh = "certmesh.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["certmesh*"]

[tool.setuptools.package-data]
certmesh = ["py.typed"]

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py310"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "SIM",  # flake8-simplify
    "TC",   # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = [
    "E501",   # line length handled by formatter
    "SIM108", # ternary not always clearer
    "SIM117", # nested with statements can be clearer in tests
    "TC001",  # move application import into type-checking block
    "TC002",  # move third-party import into type-checking block
    "TC003",  # move stdlib import into type-checking block
    "RUF043", # raw strings in match patterns - overly strict for tests
    "B008",   # Depends() in function defaults is standard FastAPI pattern
]

[tool.ruff.lint.isort]
known-first-party = ["certmesh"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["certmesh"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
